+++ python_modules/dagster/dagster_tests/core_tests/definitions_tests/test_decorators.py
@@ -162,11 +162,11 @@ def test_solid_with_explicit_empty_outputs():
         result = execute_single_solid_in_isolation(create_test_context(), hello_world)
 
     assert (
-        'Solid unexpectedly returned output foo of type <class \'str\'>. Solid is explicitly '
-        'defined to return no results.'
+        'Error in solid hello_world: Unexpectedly returned output foo of type '
+        '<class \'str\'>. Solid is explicitly defined to return no results.'
     ) in str(exc_info.value) or (
-        'Solid unexpectedly returned output foo of type <type \'str\'>. Solid is explicitly '
-        'defined to return no results.'
+        'Error in solid hello_world: Unexpectedly returned output foo of type '
+        '<type \'str\'>. Solid is explicitly defined to return no results.'
     ) in str(
         exc_info.value
     )  # py2