+++ python_modules/libraries/dagster-aws/dagster_aws/__init__.py
@@ -1,6 +1,1 @@
 from .version import __version__
-
-from .solids import download_from_s3, EmrRunJobFlowSolidDefinition
-
-
-__all__ = ['download_from_s3', 'EmrRunJobFlowSolidDefinition']