+++ python_modules/libraries/dagster-aws/dagster_aws/emr/types.py
@@ -2,19 +2,6 @@ from enum import Enum as PyEnum
 
 from dagster import Enum, EnumValue
 
-from dagster.core.types.runtime import Stringish
-from dagster.utils import safe_isfile
-
-
-class FileExistsAtPath(Stringish):
-    def __init__(self):
-        super(FileExistsAtPath, self).__init__(description='A path at which a file actually exists')
-
-    def coerce_runtime_value(self, value):
-        value = super(FileExistsAtPath, self).coerce_runtime_value(value)
-        return self.throw_if_false(safe_isfile, value)
-
-
 EbsVolumeType = Enum(
     name='EbsVolumeType', enum_values=[EnumValue('gp2'), EnumValue('io1'), EnumValue('standard')]
 )